+++ test/int/sql/test_storage_cache.py
@@ -1,0 +1,9 @@
+from framework.sqltester import (
+    ClusterSingleInstance,
+    sql_test_file,
+)
+
+
+@sql_test_file("storage_cache.sql")
+class TestStorageCache(ClusterSingleInstance):
+    pass